Tests/LibWeb/Text/expected/ShadowDOM/declarative-basic.txt

[object ShadowRoot]
0
null
0
[object ShadowRoot]
1
null
0
[object ShadowRoot]
0



<template shadowrootmode="open">open</template>


<template shadowrootmode="open" shadowrootserializable="">open</template>





<template shadowrootmode="open" shadowrootserializable="">open</template>


manual
null
<template shadowrootmode="open" shadowrootserializable="" shadowrootslotassignment="manual"><slot name="slot"></slot></template>

<span slot="slot">manual</span>

named
manual
named
named
